+++ src/main/java/nextstep/ladder/vo/Height.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.vo;
 
 public class Height {
     private final int height;
